check instance in TextListener via constructor.name
This fixes a problem which I encountered in hubot-irc (related issue
nandub/hubot-irc#2). I guess the problem is that the Robot.TextMessage object
is created in a different module and therefore the `instanceof` call fails
although it is technically a proper instance. However checking the instance
via the `constructor.name` property works.