Update scripting.md script package sample
Update the sample code for "Creating A Script Package" to use
`robot.load`. Previously it was duplicating the `exists`/`readdir` logic
that `robot.load` already has, except it was doing the exists check
asynchronously, which changes script load order and also means the
script won't have loaded in time when using the `--config-check` flag.