Fix #862 - Sorts Help Commands by Script
This changes @commands from [] → {} which keeps track of each script
as they are loaded. Then `helpCommands` returns an array after calling
sort() on the commands of each script, rather than all scripts at once.

Note that this means that the command `hubot help <TERM>` will show
commands sorted by script still, since the filtering is done in the help
script. A simple solution would be to have the help scripts call sort()
after filtering, but this, of course, wouldn't be applied to existing
installations.