Replace express.bodyParser with express.urlencoded() and express.json()
This gets rid of the two warnings reported in github/hubot/828
express.bodyParser() is deprecated in favor of only using the parts you need.
I've replaced it with only the parsing bits that I believe Hubot is using: url encoded and json.
If Hubot does in fact need to parse multipart uploads, this patch will not work.