Make bin/hubot Warn about unknown options
Previously the bin/hubot script would silently ignore options it was
passed that it didn't know about. So the following would start hubot:

    hubot --adaptor something

but wouldn't load the specified adapter as *--adaptor* not *--adapter*
was passed as an option. There was no feedback to indicate that the
option was unknown so a small mistake like this could lead to some
frustrating debugging.

This changes bin/hubot to log a warning to the console if an unknown
option is specified.