feat: Robot constructor to take an object of properties
Previously, Robot took positional arguments. This is fine for internal
use, but when programatically building, it looks weird because you have
a bunch of null and boolean values which don't make sense on their own.
It is bad for the reasons having multiple positional arguments is bad,
lack of readability.

Instead of positional arguments, this updates Robot to take an options
object, so the parameters are named and it is clear to a reader what
each does.

This continues to support the old positional arguments, but emits a
deprecation message with depd. Internal use of has been updated to use
the options object, and there probably aren't a ton of uses of Robot
directly in the wild (although there is definitely at least
hubot-test-helper). This support can be considered for removal during
the next breaking release.

This is part of https://github.com/hubotio/evolution/pull/3