fix: HUBOT_HTTPD can now disable httpd
This is a fix for https://github.com/hubotio/hubot/issues/1471. The
approach taken here is to compare the `HUBOT_HTTPD` environment variable
to the string `'false'` and then assign the value of this comparison to
the `enableHttpd` option. This ensures that a boolean value is always
passed to the robot constructor (which it already assumes).

The effect here is that setting `HUBOT_HTTPD` to `'false'` will disable
httpd, but all other values will enable it. The only backwards
compatibility issues caused by this change would be if someone happened
to already have `HUBOT_HTTPD=false` in their configuration; with this
new change they would find that httpd would actually be disabled now.