feat(robot): load adapter from local path with .mjs file extension; load scripts with .mjs and import
BREAKING CHANGE: Loading the adapter is done with async/await semantics. You can now specify an adapter with the .mjs file extension and Hubot will look in ./src/adapters for the file. In addition, scripts with the .mjs file extension can be used. They will be loaded with async/await symantics via the `import` statement.